created Quicken formula
the Quicken Formula bestows the Haste status.
right now, the Heal, Fury, and Quicken formula scripts are identical. i think i will try making a few more formulas, then try making Base scripts for each Formula Type. for example, Heal, Cleanse, Fury, and now Quicken all use an identical script, because there are no spell effects for those formulas apart from a screen tint and a sound effect. these would be of a single "Type", as in free-aimed AoE that runs a number of functions (Flash is the only unique one, because it has unique code to handle the timing of its behaviour, like the screen shake - the others are pretty much identical, other than what the formula actually does in-game). so all the formulas of a single Type will extend from a single script that governs that type, then the unique aspects of each formula's spellcast will be its own script that extends the base Type script.
the Hardball formula constitutes an entirely new type as it functions completely differently.
that's all for now.