incorporated max_speed_mod / refactored speed calculation
this stat is to modulate movement speed. right now it's used to modulate movement speed while charging. for now, max_speed is multiplied by max_speed_mod, which gets set to 0.75 while charging and 1.0 otherwise. i'm also experimenting with modulating player movement speeds via external functions. for example, map conditions, like the Desert wind.
right now the desert wind can vary in strength between the values -160 and 160; negative values imply wind blowing to the left, and vice versa. it's easy enough to set the max_speed_mod when a sand storm occurs, but it might be part of the Desert's mechanics that the wind itself actually modulates the players' positions. in this case, it might not make sense to have the wind change direction, IF the Desert is an open area that you traverse from one side to the other. if it becomes less open at some point, ie. you have to navigate traversible paths that change direction, like a desert maze, then it makes sense to have the wind change direction, as well as the degree to which the players' positions get modulated by the wind, in either direction.
i want the desert map to be challenging to traverse. if the wind is at your back, ie. you're being blown towards your destination, that kind of defeats the theme of the map. just something to consider while thinking through these issues.
also as a test i have max_speed_mod emit a signal recieved by the Player to set its Run animation timescale to the value of max_speed_mod. however, i think it doesn't look great (because the Run animation doesn't have a lot of frames) so i think i will have to create a separate movement animation for charging Players, similar to secret of mana, perhaps with sword drawn.
also got rid of the vestiges of the old status "system". the only thing it was being used for was to detect "sweating" and making the GUI's StaminaBar flash, which is now useless because the ChargeUI flashes when low on stamina, and sweating goes away when stamina is above 0 now, unlike before, when it was possible to be sweating with stamina. i've kept the signals emitted from PlayerStats (but renamed them from "status" to "sweating") just in case i still want to do something with it. but for now, the effect is basically unnoticeable.