adjusted Camera2D behaviour for multiplayer
needs a lot of tuning, but now the player that isn't moving gets teleported to the one that is, and vice versa, as opposed to P2 simply teleporting to P1's position.
uses a lot of dirty distance checking, so it doesn't "feel" good yet.
also adjusting some objects placements on some of the maps, and connecting some more test maps together.