level up screen no longer instantiates automatically
i'm still testing how i want to do this so it's not fully implemented, yet.
level ups occur as normal when experience exceeds experience_required, but we do not start the "LevelTimer" of the levelling player anymore. instead, we increase a new stat, "attributes_to_allocate".
right now, the test key "5" starts the LevelTimers. when it times out, we check for stats.attributes_to_allocate, and instantiate a Level Up Screen if it's more than 1.
when we click a stat button, we decrease attributes_to_allocate  by 1.
have to come up with a way to cancel out of the level Up screen, i think.