current selected formula remains on scene change / pixel perfect applied / sound fix
the pixel perfect shader makes sure pixels are rendered on the resolution grid of the viewport, though this comes with a number of issues, mainly that the game feels less smooth. will have to modify it as we go along.
the formulabooks get instantiated with a default current_selected_formula value of 0, which always set the selected formula to the first one in the formulas array. now, when we reinitialize the players, we grab the previous current_selected_formula value from the inventory reference created by the Exit. then we call the new set_selected_formula(formula_index) method, which emits the signal heard by the FormulaUI, setting it to the appropriate formula selected.
i also realized a redundancy between the intialize_player() method and the reinitialize_player() method, and think that only one function is needed. the reinitialize_player() method was called when a Player changed scenes, creating an inventory reference, while the initialize_player() was called when a Player was created and an inventory reference did not exist. now, the initialize_player method checks for null values on p#_pouch and p#_formulabook, which are the variables used to store either resources that have been loaded from disk OR empty resources created when selecting "new game". after we set the inventory, we set the p# variables to null, ensuring resources don't get double loaded when players get initialized again. curiously, when they were double loading, the correct inventory was set even if we attempted to load from the resource placeholder variables. i'm can't understand why.  anyways i hope i don't run into issues with this, because the multiplayer/loading/saving/inventory aspect is one of the most complex things in the entire program and it's really difficult to grasp the flow of code and what's actually happening in what sequence.
i've also run into an engine issue regarding pausing the sceneTree and AudioStreamPlayers:
if the music gets paused - for example, when dying - then, the player does something that pauses the SceneTree - for example, casting a formula - then when the SceneTree unpauses, it automatically unpauses any AudioStreamPlayer children as well. this is highly undesirable, but it confirms a previous issue i was having where ending a LevelUpScreen would also un-pause any paused AudioStreams, because it was resuming the SceneTree.
so a number of dirty fixes had to be implemented. first, you can't pause the game when dying. i'm fine with this. the dirtier trick is when we end the FormulaTargetScreen:
after we end the target screen, the SceneTree unpauses (which unpauses the music if we are dying, which we don't want). we then iterate through the nodes in group "Players", and check if any's dying variable = true, and if it is, we RE-PAUSE the music, and set a variable - "dying_effect" - to true. then we check for !dying_effect, and remove the Low Pass filter on the audio if dying_effect = false. we need to do this because the Dying effect enables a Low Pass filter on the Master audio channel, as does dying, but we don't want to disable the filter when we exit the formula screen while dying. i hope the Pause system in Godot gets updates because it's very unwieldy. there is no documentation anywhere that says AudioStreams get unpaused when the SceneTree gets unpaused. more confusing, AudioStreams don't automatically get paused when the SceneTree is paused, unless their PAUSE_MODE = "Stop".