working on multiplayer controls for menus
one of the most annoying things when playing locally is when both players can control the same thing, like a user interface. i am going to write a system that sets the controls of a menu to be appropriate to the player that opened the menu. this includes the pause menu, the formula target screen, and the level up screen.
i hope to have each of these screens declare a dictionary of empty input keys. then, we run a Global.set_player_menu_controls() function, which populates the dictionary keys with values corresponding to the appropriate player's controls. i see this as a re-usable way to set menu controls, and it should be scaleable for more than 2 players.
right now, only the FormulaTargetScreen is hard-coded properly. i will clean up and make sure the controls setup for each menu is uniform so as to avoid any unexpected issues.