controls dict setup working for FormulaTargetScreen
depending on who casts, the controls get set properly.
unfortunately, this is not easy to do for the Pause menu and the Level up screen because navigating those involves control nodes and shifting focus, which is done via the input events that begin with "ui_", which are built-in to the engine. clearly, the Input system was not designed with multiplayer in mind.
i guess it might be possible to interpret player input events during menus, then depending on the Player inputting, either allowing the input event to propogate or not. i just can't think of a way of doing that cleanly.
right now the menus are coded very haphazardly.
i have had to hard-code the Pause Menu to ignore input events that correspond with the player that DID NOT open the pause screen, which is absurd. the same with the level Up screen.
even worse, these hard-coded keys to ignore do not include whether a modifier key like ALT or SHIFT is being held when the key is struck. meaning, Player1 can "confirm" Player2's buttons by holding SHIFT and pressing their attack key. it's fucking absurd, and it's fucking terrible, and i can't do fucking shit about it because the Engine is built this way. it's fucking absurdly fucking bad. it's fucking horrible.
why are control nodes inherently tied to "actions" that begin with "ui_"? why is there no way to stop that from happening? it's so fucking badly designed. it's such a colossal waste of time to have to deal with this. having to program separate input handlers JUST TO IGNORE EVERY KEY BUT THE ONES YOU WANT.
the cleanest way i can think of doing this, and it's fucking stupid, and fucking overly complex for something that should be stupidly fucking easy, is to have every "ui_" action empty by default, then assign keys to them when menus are opened, then to make them empty again when the menus disappear. i'm assuming InputMap events can even be set via code, which i don't know if they can. what a fucking nightmare and a bunch of time is going to be spent on this. fuck menus, fuck UI, fuck input events. the amount of unpredictable/straight-up fucking terrible behaviour out of these systems is mind-boggling. i cannot believe anybody can reasonably put up with how fucking awfully designed this entire mess is.